+++ src/utils/defaultMediaFolder.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_MEDIA_FOLDER = "LumaFrame Media";